ai/requirements_Anaconda.txt
absl-py==2.2.2
addict==2.4.0
aioboto3==14.3.0
aiobotocore==2.22.0
aiofiles==24.1.0
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aioitertools==0.12.0
aiosignal==1.3.1
alabaster==0.7.13
albucore==0.0.17
albumentations==1.2.1
aliyun-python-sdk-core==2.16.0
aliyun-python-sdk-kms==2.16.5
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==25.3.0
audioread==3.0.1
av==12.3.0
babel==2.17.0
bcrypt==4.0.1
boto3==1.37.3
botocore==1.37.3
CacheControl==0.14.2
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
chumpy==0.70
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.1.1
coverage==7.6.1
crcmod==1.7
cryptography==44.0.3
cycler==0.12.1
Cython==3.1.0
decorator==4.4.2
decord==0.6.0
deep-sort-realtime==1.3.2
docker-pycreds==0.4.0
docutils==0.18.1
easydict==1.13
einops==0.8.1
eval_type_backport==0.2.2
exceptiongroup==1.2.2
fastapi==0.95.2
filelock==3.14.0
firebase-admin==6.8.0
flake8==7.1.2
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2024.6.1
ftfy==6.2.3
future==1.0.0
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.1.0
google-api-core==2.24.2
google-api-python-client==2.169.0
google-auth==2.40.1
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.0.0
google-cloud-core==2.4.3
google-cloud-firestore==2.20.2
google-cloud-storage==3.1.0
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpcio==1.70.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httpx==0.25.1
huggingface-hub==0.31.1
humanfriendly==10.0
idna==3.10
imageio==2.35.1
imageio-ffmpeg==0.5.1
imagesize==1.4.1
imgaug==0.4.0
importlib_metadata==8.5.0
importlib_resources==6.4.5
iniconfig==2.1.0
insightface==0.7.3
intel-openmp==2021.4.0
interrogate==1.7.0
isort==4.3.21
Jinja2==3.1.4
jmespath==0.10.0
joblib==1.4.2
json-tricks==3.17.3
kiwisolver==1.4.7
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.41.1
lmdb==1.6.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mccabe==0.7.0
mdit-py-plugins==0.4.2
mdurl==0.1.2
mkl==2021.4.0
-e git+https://github.com/open-mmlab/mmaction2.git@4d6c93474730cad2f25e51109adcf96824efc7a3#egg=mmaction2
#mmcv==2.1.0
mmdet==3.2.0
mmengine==0.10.4
mmpose==1.3.2
model-index==0.1.11
modelindex==0.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
munkres==1.1.4
myst-parser==3.0.1
networkx==3.1
numba==0.58.1
numpy==1.24.4
oauthlib==3.2.2
onnx==1.17.0
onnxruntime-gpu==1.15.1
openai-clip==1.0.1
openai-whisper==20240930
opencv-contrib-python==4.11.0.86
opencv-python==4.5.5.64
opencv-python-headless==4.11.0.86
opendatalab==0.0.10
openmim==0.3.9
openxlab==0.1.2
ordered-set==4.1.0
oss2==2.17.0
packaging==24.0
pandas==2.0.3
parameterized==0.9.0
passlib==1.7.4
pillow==10.4.0
PIMS==0.7
platformdirs==4.3.6
pluggy==1.5.0
pooch==1.8.2
prettytable==3.11.0
proglog==0.1.12
propcache==0.2.0
proto-plus==1.26.1
protobuf==4.23.4
psutil==7.0.0
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycocotools==2.0.7
pycodestyle==2.12.1
pycparser==2.22
pycryptodome==3.22.0
pydantic==1.10.14
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.1.4
pyreadline3==3.5.4
pytest==8.3.5
pytest-runner==6.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.6
PyTurboJPEG==1.7.7
pytz==2023.4
PyWavelets==1.4.1
#pywin32==310
PyYAML==6.0.2
qudida==0.0.4
regex==2024.11.6
requests==2.28.2
requests-oauthlib==2.0.0
rich==13.4.2
rsa==4.9.1
s3transfer==0.11.3
safetensors==0.5.3
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
sentry-sdk==2.27.0
setproctitle==1.3.6
shapely==2.0.7
six==1.17.0
slicerator==1.1.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
soundfile==0.13.1
soxr==0.3.7
Sphinx==6.1.3
sphinx-copybutton==0.5.2
sphinx-markdown-tables==0.0.17
sphinx-notfound-page==1.1.0
sphinx-tabs==3.4.7
sphinxcontrib-applehelp==1.0.4
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.1
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
starlette==0.27.0
sympy==1.13.3
tabulate==0.9.0
tbb==2021.11.0
tensorboard==2.14.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminaltables==3.1.10
threadpoolctl==3.5.0
tifffile==2023.7.10
tiktoken==0.7.0
timm==1.0.15
titlecase==2.4.1
tokenizers==0.20.3
tomli==2.2.1
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tqdm==4.65.0
transformers==4.46.3
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==1.26.20
uvicorn==0.23.2
wandb==0.19.11
wcwidth==0.2.13
webrtcvad==2.0.10
Werkzeug==3.0.6
wrapt==1.17.2
xdoctest==1.2.0
xtcocotools==1.14.3
yapf==0.43.0
yarl==1.15.2
zipp==3.20.2
azure-storage-blob==12.25.1
